requirement.txt
asgiref==3.5.2
asttokens==2.0.5
attrs==22.2.0
autopep8==2.0.1
backcall==0.2.0
black @ git+https://github.com/psf/black@6407ebb870afe0062ee581abdea07c1ef5213d31
click==8.1.3
colorama==0.4.6
decorator==5.1.1
Django==4.1.5
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
executing==0.9.1
inflection==0.5.1
iniconfig==2.0.0
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
packaging==23.0
parso==0.8.3
pathspec==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.2
pluggy==1.0.0
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.12.0
pytest==7.2.1
pytest-django==4.5.2
pytest-testdox==3.0.1
pytz==2022.6
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
traitlets==5.3.0
tzdata==2022.7
wcwidth==0.2.5
